refatorações e solução de bugs

+++ backend/src/main/java/br/com/budgetflow/features/categorias/service/CategoriaService.java
@@ -2,7 +2,9 @@ package br.com.budgetflow.features.categorias.service;
 
 import br.com.budgetflow.common.enums.ClassificacaoCategoria;
 import br.com.budgetflow.common.exceptions.ConflictException;
+import br.com.budgetflow.common.exceptions.EntityHasRelationshipsException;
 import br.com.budgetflow.common.exceptions.ResourceNotFoundException;
+import br.com.budgetflow.common.service.RelacionamentoChecker;
 import br.com.budgetflow.features.categorias.domain.Categoria;
 import br.com.budgetflow.features.categorias.dto.CategoriaRequestDTO;
 import br.com.budgetflow.features.categorias.dto.CategoriaResponseDTO;
@@ -25,11 +27,13 @@ public class CategoriaService {
 	private final CategoriaRepository categoriaRepository;
 	private final UserService userService;
 	private final CategoriaMapper categoriaMapper;
+	private final RelacionamentoChecker relacionamentoChecker;
 
-	public CategoriaService(CategoriaRepository categoriaRepository, UserService userService, CategoriaMapper categoriaMapper) {
+	public CategoriaService(CategoriaRepository categoriaRepository, UserService userService, CategoriaMapper categoriaMapper, RelacionamentoChecker relacionamentoChecker) {
 		this.categoriaRepository = categoriaRepository;
 		this.userService = userService;
 		this.categoriaMapper = categoriaMapper;
+		this.relacionamentoChecker = relacionamentoChecker;
 	}
 
 	@Transactional
@@ -96,6 +100,13 @@ public class CategoriaService {
 	public void delete(Long id) {
 		Long userId = SecurityUtils.currentUserId();
 		Categoria categoria = findByIdAndUser(id, userId);
+
+		if (relacionamentoChecker.categoriaHasRelationships(id, userId)) {
+			throw new EntityHasRelationshipsException(
+					"A categoria \"" + categoria.getNome() + "\" não pode ser excluída pois está vinculada a uma ou mais transações"
+			);
+		}
+
 		categoriaRepository.delete(categoria);
 	}
